fix: keep forked child process alive via SIGTERM #287
When exporting an array of configs from webpack.config.js, atl crashes.
The error message is "channel closed", indicating that a forked child
process is dying too soon. This commit changes the Checker class such
that it's kill method sends a SIGTERM to the child process rather
than a SIGKILL. The child process intercepts this signal and
ignores it to ensure it is kept alive.

Closes #287